Split schema migrations into migration package
