+++ internal/repository/migration/migration.go
@@ -1,0 +1,94 @@
+package migration
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/sirupsen/logrus"
+	"gorm.io/gorm"
+)
+
+const (
+	migrationAddUsageEventRedisFields               = "20260503_add_usage_event_redis_fields"
+	migrationBackfillUsageEventRedisFields          = "20260503_backfill_usage_event_redis_fields"
+	migrationDropSnapshotRuns                       = "20260503_drop_snapshot_runs"
+	migrationDropLegacySnapshotRunColumns           = "20260504_drop_legacy_snapshot_run_columns"
+	migrationCreateUsageIdentities                  = "20260504_create_usage_identities"
+	migrationMigrateUsageIdentitiesMetadata         = "20260504_migrate_usage_identities_metadata"
+	migrationBackfillUsageEventIdentityFields       = "20260504_backfill_usage_event_identity_fields"
+	migrationBackfillUsageIdentityStats             = "20260504_backfill_usage_identity_stats"
+	migrationDropLegacyMetadataTables               = "20260504_drop_legacy_metadata_tables"
+	migrationRemovePrefixUsageIdentities            = "20260504_remove_prefix_usage_identities"
+	migrationAddUsageIdentityLookupKey              = "20260505_add_usage_identity_lookup_key"
+	migrationMigrateAIProviderIdentitiesToAuthIndex = "20260505_migrate_ai_provider_identities_to_auth_index"
+)
+
+type schemaMigration struct {
+	Version   string    `gorm:"primaryKey;column:version"`
+	AppliedAt time.Time `gorm:"not null;column:applied_at"`
+}
+
+func (schemaMigration) TableName() string {
+	return "schema_migrations"
+}
+
+type databaseMigration struct {
+	version string
+	run     func(*gorm.DB) error
+}
+
+func Run(db *gorm.DB) error {
+	if err := db.Exec("CREATE TABLE IF NOT EXISTS schema_migrations (version TEXT PRIMARY KEY, applied_at DATETIME NOT NULL)").Error; err != nil {
+		return fmt.Errorf("create schema_migrations table: %w", err)
+	}
+
+	for _, migration := range orderedMigrations() {
+		if err := runSchemaMigration(db, migration); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func orderedMigrations() []databaseMigration {
+	return []databaseMigration{
+		{version: migrationAddUsageEventRedisFields, run: addUsageEventRedisFieldsMigration},
+		{version: migrationBackfillUsageEventRedisFields, run: backfillUsageEventRedisFieldsMigration},
+		{version: migrationDropSnapshotRuns, run: dropSnapshotRunsMigration},
+		{version: migrationDropLegacySnapshotRunColumns, run: dropLegacySnapshotRunColumnsMigration},
+		{version: migrationCreateUsageIdentities, run: createUsageIdentitiesMigration},
+		{version: migrationMigrateUsageIdentitiesMetadata, run: migrateUsageIdentitiesMetadataMigration},
+		{version: migrationBackfillUsageEventIdentityFields, run: backfillUsageEventIdentityFieldsMigration},
+		{version: migrationBackfillUsageIdentityStats, run: backfillUsageIdentityStatsMigration},
+		{version: migrationDropLegacyMetadataTables, run: dropLegacyMetadataTablesMigration},
+		{version: migrationRemovePrefixUsageIdentities, run: removePrefixUsageIdentitiesMigration},
+		{version: migrationAddUsageIdentityLookupKey, run: addUsageIdentityLookupKeyMigration},
+		{version: migrationMigrateAIProviderIdentitiesToAuthIndex, run: migrateAIProviderIdentitiesToAuthIndexMigration},
+	}
+}
+
+func runSchemaMigration(db *gorm.DB, migration databaseMigration) error {
+	return db.Transaction(func(tx *gorm.DB) error {
+		logger := logrus.WithField("version", migration.version)
+		var count int64
+		if err := tx.Table("schema_migrations").Where("version = ?", migration.version).Count(&count).Error; err != nil {
+			logger.WithError(err).Error("schema migration failed")
+			return fmt.Errorf("check schema migration %s: %w", migration.version, err)
+		}
+		if count > 0 {
+			logger.Info("schema migration skipped")
+			return nil
+		}
+		logger.Info("schema migration started")
+		if err := migration.run(tx); err != nil {
+			logger.WithError(err).Error("schema migration failed")
+			return fmt.Errorf("run schema migration %s: %w", migration.version, err)
+		}
+		if err := tx.Create(&schemaMigration{Version: migration.version, AppliedAt: time.Now().UTC()}).Error; err != nil {
+			logger.WithError(err).Error("schema migration failed")
+			return fmt.Errorf("record schema migration %s: %w", migration.version, err)
+		}
+		logger.Info("schema migration applied")
+		return nil
+	})
+}
